+++ src/frontend/apps/drive/src/features/auth/components/LogoutButton.tsx
@@ -1,9 +1,11 @@
 import { Button } from "@openfun/cunningham-react";
-import { logout } from "../Auth";
 import { useTranslation } from "react-i18next";
+import { useAuth } from "../Auth";
 
 export const LogoutButton = () => {
   const { t } = useTranslation();
+  const { logout } = useAuth();
+
   return (
     <Button variant="tertiary" onClick={logout} fullWidth={true}>
       {t("logout")}