+++ src/frontend/apps/drive/src/hooks/useThemeCustomization.tsx
@@ -1,13 +1,12 @@
 import { useConfig } from "@/features/config/ConfigProvider";
-import { ThemeCustomization } from "@/features/drivers/types";
 import { splitLocaleCode } from "@/features/i18n/utils";
 import { useTranslation } from "react-i18next";
 
-export const useThemeCustomization = (key: keyof ThemeCustomization) => {
+export const useThemeCustomizationFooter = () => {
   const { config } = useConfig();
   const { i18n } = useTranslation();
   const language = splitLocaleCode(i18n.language).language;
-  const themeCustomization = config?.theme_customization?.[key];
+  const themeCustomization = config?.theme_customization?.footer;
   return {
     ...themeCustomization?.default,
     ...(themeCustomization?.[language as keyof typeof themeCustomization] ??