+++ src/frontend/apps/drive/src/pages/401.tsx
@@ -1,4 +1,4 @@
-import { login } from "@/features/auth/Auth";
+import { useAuth } from "@/features/auth/Auth";
 import { getSimpleLayout } from "@/features/layouts/components/simple/SimpleLayout";
 import { GenericDisclaimer } from "@/features/ui/components/generic-disclaimer/GenericDisclaimer";
 import { Button } from "@openfun/cunningham-react";
@@ -6,12 +6,14 @@ import { useTranslation } from "react-i18next";
 
 export default function UnauthorizedPage() {
   const { t } = useTranslation();
+  const { login } = useAuth();
+
   return (
     <GenericDisclaimer
       message={t("401.title")}
       imageSrc="/assets/401-background.png"
     >
-      <Button onClick={() => login()}>{t("401.button")}</Button>
+      <Button onClick={() => login?.()}>{t("401.button")}</Button>
     </GenericDisclaimer>
   );
 }