+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- ✨(front) allow homepage redirection through config
+
+### Changed
+
+- ♻️(front) move auth urls to driver for config
+
 ## [v0.10.1] - 2025-12-05
 
 ### Security